* MERICA!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
Now you have to answer to
 A  M  E  R  I  C  A !  !! !!! !!  - Team America

This is the last of the language normalization (not normalisation)
efforts out of my tree, also fixes spelling errors (unrelated to
British vs. American English) in some comments and documentation.

There is now a **single** British English word remaining, and it
happens to be in the *seedwords*, of all places! It's **"artefact"**,
which is definitively British! So weird. Anyway, I left that alone,
because I assume changing the seedwords now would have consequences
for existing seeds, so I'll have to live with "artefact", I guess.