* Improvements for: peer, syncmgr, do, pktwallet.
* netsync/manager:
 1. faster stall detection
 2. attempt to avoid ban triggers

* peer/peer:
 1. additional locking
 2. only increase block height
 3. increase `DefaultTrickleInterval` time
 4. allow (and log) unhandled commands

pktwallet/config:
 1. ensure data directory creation

do: 1. extensive additional error checking
    2. enhanced POSIX compliance **

* ** Explanation and rationale for `do` changes:

1. Switch to `env` to locate shell - `$SHELL`
is nonstandard (and there is no standard
shell location in POSIX); you can use the
`command -p getconf _POSIX_SHELL` command
to check if a POSIX-compatible shell is
supplied, but only `env` has a standarized
location, so it can be used to find `sh`.

2. Strange use of `true :` is a workaround
for known shell bugs and a noop elsewhere.

3. `set -e` error hanlding is buggy on **most**
shells - it's behavior varies so widely
that it cannot be relied on, especially
in combination with traps - manual error
checking via `try()`/`die()` is a common and
and reliable workaround, and the `just()`
function allows a script to work the same
both with and without -e error handling.

4. `echo` must never be used and is nonstandard.

"According to POSIX, `echo` has unspecified
behavior if any of its arguments contain `“\”`
or if it's first argument is “-n”. Various UNIX™
"standards" fill in this unspecified area for
XSI-conformant implementations, by specifying
nasty undesirable behavior that no one wants:
(`“\”` is interpreted as a C-string-literal
style escape, for example), and other popular
"standard" implementations such as Bash will
interpret argument values other than `“-n”`
as special options, even when with the dubious
*“POSIX compatibility”* mode enabled, rendering
them non-conformant - ksh offers `"print"`
which is very sane, but sadly, `ksh`-specific.

Due to these limitations, `echo` can only be
reliably used in a very small number of cases,
and only where you have strict guarantees about
the contents of any variable being evaluated,
including oddness that's easy to forget about,
like the fact that the variable may not contain
any non-negative integer, etc.

`say()` wrap's `printf` and avoids this idiocy.

5. All variables are modified via POSIX variable
substutiton to be either defaulted or checked,
making the script compatible with shells that
default or enable the "set -u" option which
modifies the checking in the shell's expansion
rules.

6. Some rather pedantic checks are enabled, such
as checking the return levels for "export", as
it may fail due to lack of environment space.

7. A wrapper is defined to use `command` whenever
we can, to suppress the behavior of the POSIX
lookup function 1.b. Command Search and
Execution rules, to normalize error handling.

8. Where available (such as BSD or Linux systems),
the "stdbuf" utility, and "--line-buffered"
argument to grep are enabled automatically,
allowing "do" to specially disable output
buffering, which provides slightly slower,
but immediate feedback when filering output.

9. `git diff --ignore-cr-at-eol` is used, so as
to not fail when the source code is modified
between Windows and UNIX systems, for example,

10. Since we now require Go 1.14+, we can use the
automatic linker selection and trimpath to
get closer to and even actually create (sometimes)
reproducible binaries, on some, but not
all platforms, using the internal Go linker
where possible (while still passing "-static"
to the external linker if selected), and
requesting builds with CGo disabled. If your
on a platform where Go would use the external
linker, and that linker is GNU, then you do
require additional flags to make reproducible
binaries, such as disabling generation of the
GNU-Build-ID and other tweaks. This is **NOT**
done currently, but will be available when the
`magefile`-based build system lands.