* Last British -> American + spelling correction
This is the last of the language normalization
(**not normalisation**) efforts out of my tree,
which also fixes an egregious number of spelling
errors completely unrelated to to British vs.
American English spelling differences, in code,
comments, and documentation.

There is now a **single** British English
word remaining, and it happens to be in the
*seedwords*, of all places!

It's **"artefact"**, which is *definitively*
Brish English! The only one. Very *weird*.

Maybe we should add "`lorry`", and "`trollop`",
and "`blimey`", and "`knackered`" in there as
well, just to even it out some?

(**Nooooooooooooooooooooo!!!!**)

Anyway, I left that alone, because I confirmed
changing the seedwords now would have
consequences for existing seeds, so we'll have
to live with the *"artefact"* forever, I guess.

(RT uses "Requestor" instead of "Requester", yet
I've *somehow* still been able to live my life.)