do: Avoid implicit @ expansion, use POSIX command
TL;DR - yeah you actually do want this change.

Why?

This deserves a real explanation on why this is the more 'correct',
or, perhaps, "desired", is the proper term, for `$*` vs. `$@` ...

First off, `$@` (`${@}`) is unreliable as it is not actually uniformly
handled across all shells in the unquoted form (vs. `"${@}"`), while
the handling of `$*` and `"$*"` is, in real life usages, deterministic.

When using the quoted version, it cannot be implicitly concatenated.
For example, while working on most shells, trying to something like
`printf '%s\n' "stuff $@"` puts you at the mercy of pure undefined and
unspecified behaviors!  While proper static analysis tools such as SC,
IntelliJ, CDT, etc. will treat that as an error, real life varies.

Now, in your case, while you aren't using that form, and, apparently you
don't want the quoted form (which will not perform IFS expansions), so
you *do* want shell expansion to take place here, the only actual real
difference here is the fact that `${*}`/`$*`, when expanded, will be subject
to both shell expansion per POSIX IFS rules *as well as* globbing. For this
sort of use case, the globbing is not necessarily unwanted, as input
will be both simple and mostly controlled, so the `$*` format is exactly
equal to the `$@` format but will be more widely compatible!

Best practices for POSIX portability strongly encourage you to always
pass all position parameters in the exact way that they are passed in,
that is, the exact form as set initially, so, `"$@"`, (in the explicitly
quoted form) is what you almost always want, especially in bash, ksh, and
zsh ... but, you can't always rely on the sh you actually get being one of
those shells, especially considering the long list of supported Go systems.

From the `bash` and AT&T ksh documentation:
('c' here represents the first IFS character defined.)

+------+-------------------------+
| Form |               Expansion |
+------+-------------------------+
|  $*  | $1 $2 $3 … ${N}         |
|  $@  | $1 $2 $3 … ${N}         |
| "$*" | "$1c$2c$3c…c${N}"       |
| "$@" | "$1" "$2" "$3" … "${N}" |
+------+-------------------------+

(Side note: It would be even better would be to use explicit substring
expansion, but, this is also not supported uniformly across all shells.)

Now, the only major consideration is to never use $* when **processing**
most user or other untrusted inputs, due to the globbing... for example,
imagine a case where files are maliciously named, such as a file called
`"rm "` and another one called `"-rf ."`. If you are looping, eval'ing,
or otherwise processing `$*`, the expansion would be to `"rm -rf ."`!
Also, `*` or other globs would be expanded as well if not escaped. With
`'${@}'`, it would just expand to one `"rm "` arg and one `"-rf ."` arg.
While this isn't **directly** applicable here, it's well worth noting.

Since you want to keep the shell side of this short, and avoid doing
extra shell-based checking, which would need to be done in a strictly
POSIX or SUSv3 compliant way, using `$*` rather than `$@ ` will give you
results that are *EXACTLY* equal - except `$*`/`${*}` is more portable.

This is my rationale and justification for for the change. Now, I could
use the best form, which would depend on the shell in use, so to do so
(in a fully POSIX compliant way) means adding code - probably more than
10 lines of new code here, just for this, probably *not* what you want.

Anyway, this is all going to be replaced with Mage, maybe by the end of
my work with this branch today, so, hopefully, this is all a short-lived
fix - but it was totally triggering me!

Anyway, rule of thumb: `"${*} over ${@} always! (for the unquoted form)"`,
but, when not using the unquoted form, `"${@}" over "${*}"` in every case!
This way, you're most likely actually doing what you really intended to do.

Now ... for the use of `command`, it makes sense to do this here because
you already are using the POSIX complaint method to determine the shell,
by way of `env`, which is the best practice, except, you can't know, at
least not without dozens and dozens of lines of code, which shell you is
going to end up executing the script. Also, you can't know if this shell
that was called has done things such as implicitly imported new functions
(hello bash!) or performed other initialization steps that we aren't privy
to and don't want. Rather than add more lines handling these potential
special cases, the easiest way to ensure we are calling "go" the program
(and not "go" the function, or "go" the alias, or "go" otherwise molested)
is by the use of the `command` command, which is universally supported,
and, being part of very early POSIX, can only add compatibility, so there
is zero downside!  Use of `builtin` would be even better for this use, but
that's a `ksh`-ism (but one that happens to be supported in `zsh` and `bash`
also), so we can't use it, if we want to be compatible across all Go systems.

Practical example #1:           (FYI using ksh, so pardon my `whence`ing)
» alias go="false"
» whence go
← false
» go version; printf '%s\n' "${?}"
← 1
» command go version; printf '%s\n' "${?}"
← go version devel +5a4db102b2 Thu Dec 17 03:20:55 2020 +0000 linux/amd64
← 0

Practical example #2 (continued from #1)"
» unalias go
» whence go
← /opt/go/bin/go
» go () { false "${@}" }
» whence go
← go
» go version; printf '%s\n' "${?}"
← 1
» command go version; printf '%s\n' "${?}"
← go version devel +5a4db102b2 Thu Dec 17 03:20:55 2020 +0000 linux/amd64
← 0

`bash` example, automatic importation of exported functions (aka idiocy):
(Note that bash was actually exec'd here in a clean environment and no -E)

» exec env bash
$ printf '%s\n' "${BASH_VERSINFO[*]}"               # yup, bash
← 5 0 17 1 release x86_64-redhat-linux-gnu
$ function go() { false "${@}"; }; export -f go     # bash idiocy
$ exec env bash                                     # new shell!
$ which go
← /opt/go/bin/go                                    # a lie!
$ go version; printf '%s\n' "${?}"
← 1                                                 # foiled!
$ declare -x -f                                     # mmmmm
← go ()
← {
←     false "${@}"                                  # sneaky! sneaky!
← }; declare -fx go
$ command go version; printf '%s\n' "${?}"
← go version devel +5a4db102b2 Thu Dec 17 03:20:55 2020 +0000 linux/amd64
← 0

We could have used a ksh-ism, `type`, to known about the imported function:

$ type go
← go is a function
← go ()
← {
←     false "${@}"
← }

... but of course, that isn't portable either, and not part of POSIX, so,
you see, in this case, using `command` will makes sure we are doing what
we expect, and that we aren't at the mercy of random outside factors, all
without adding any additional lines of code to check for all this stupidity.

:)